pyproject.toml
[project]
name = "deep-research-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "gradio>=6.2.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-groq>=1.1.1",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "langsmith>=0.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.17",
]
